requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
jinja2>=3.1.0
aiofiles>=23.0.0
itsdangerous>=2.1.0
httpx>=0.24.0
